+++ src/collect.jl
@@ -24,29 +24,29 @@ ArrayInitializer(unwrap = t->false) = ArrayInitializer(unwrap, arrayof)
 _reshape(v, itr, ::Base.HasShape) = reshape(v, axes(itr))
 _reshape(v, itr, ::Union{Base.HasLength, Base.SizeUnknown}) = v
 
-function collect_fieldarrays(itr; initializer = default_initializer)
+function collect_structarray(itr; initializer = default_initializer)
     sz = Base.IteratorSize(itr)
-    v = collect_fieldarrays(itr, sz, initializer = initializer)
+    v = collect_structarray(itr, sz, initializer = initializer)
     _reshape(v, itr, sz)
 end
 
-function collect_empty_fieldarrays(itr::T; initializer = default_initializer) where {T}
+function collect_empty_structarray(itr::T; initializer = default_initializer) where {T}
     S = Core.Compiler.return_type(first, Tuple{T})
     initializer(S, (0,))
 end
 
-function collect_fieldarrays(@nospecialize(itr), ::Union{Base.HasShape, Base.HasLength};
+function collect_structarray(@nospecialize(itr), ::Union{Base.HasShape, Base.HasLength};
     initializer = default_initializer)
 
     st = iterate(itr)
-    st === nothing && return collect_empty_fieldarrays(itr, initializer = initializer)
+    st === nothing && return collect_empty_structarray(itr, initializer = initializer)
     el, i = st
     dest = initializer(typeof(el), (length(itr),))
     dest[1] = el
-    collect_to_fieldarrays!(dest, itr, 2, i)
+    collect_to_structarray!(dest, itr, 2, i)
 end
 
-function collect_to_fieldarrays!(dest::AbstractArray{T}, itr, offs, st) where {T}
+function collect_to_structarray!(dest::AbstractArray{T}, itr, offs, st) where {T}
     # collect to dest array, checking the type of each result. if a result does not
     # match, widen the result type and re-dispatch.
     i = offs
@@ -58,24 +58,24 @@ function collect_to_fieldarrays!(dest::AbstractArray{T}, itr, offs, st) where {T
             @inbounds dest[i] = el
             i += 1
         else
-            new = widenfieldarrays(dest, i, el)
+            new = widenstructarray(dest, i, el)
             @inbounds new[i] = el
-            return collect_to_fieldarrays!(new, itr, i+1, st)
+            return collect_to_structarray!(new, itr, i+1, st)
         end
     end
     return dest
 end
 
-function collect_fieldarrays(itr, ::Base.SizeUnknown; initializer = default_initializer)
+function collect_structarray(itr, ::Base.SizeUnknown; initializer = default_initializer)
     elem = iterate(itr)
-    elem === nothing && return collect_empty_fieldarrays(itr; initializer = initializer)
+    elem === nothing && return collect_empty_structarray(itr; initializer = initializer)
     el, st = elem
     dest = initializer(typeof(el), (1,))
     dest[1] = el
-    grow_to_fieldarrays!(dest, itr, iterate(itr, st))
+    grow_to_structarray!(dest, itr, iterate(itr, st))
 end
 
-function grow_to_fieldarrays!(dest::AbstractArray{T}, itr, elem = iterate(itr)) where {T}
+function grow_to_structarray!(dest::AbstractArray{T}, itr, elem = iterate(itr)) where {T}
     # collect to dest array, checking the type of each result. if a result does not
     # match, widen the result type and re-dispatch.
     i = length(dest)+1
@@ -86,9 +86,9 @@ function grow_to_fieldarrays!(dest::AbstractArray{T}, itr, elem = iterate(itr)) 
             elem = iterate(itr, st)
             i += 1
         else
-            new = widenfieldarrays(dest, i, el)
+            new = widenstructarray(dest, i, el)
             push!(new, el)
-            return grow_to_fieldarrays!(new, itr, iterate(itr, st))
+            return grow_to_structarray!(new, itr, iterate(itr, st))
         end
     end
     return dest
@@ -99,10 +99,10 @@ function to_structarray(::Type{T}, nt::C) where {T, C}
     StructArray{S}(nt)
 end
 
-function widenfieldarrays(dest::StructArray{T}, i, el::S) where {T, S}
+function widenstructarray(dest::StructArray{T}, i, el::S) where {T, S}
     fs = fields(S)
     if fs === fields(T)
-        new_cols = (widenfieldarrays(fieldarrays(dest)[ind], i, getfieldindex(el, f, ind)) for (ind, f) in enumerate(fs))
+        new_cols = (widenstructarray(fieldarrays(dest)[ind], i, getfieldindex(el, f, ind)) for (ind, f) in enumerate(fs))
         nt = NamedTuple{fs}(Tuple(new_cols))
         v = to_structarray(T, nt)
     else
@@ -110,7 +110,7 @@ function widenfieldarrays(dest::StructArray{T}, i, el::S) where {T, S}
     end
 end
 
-widenfieldarrays(dest::AbstractArray, i, el) = widenarray(dest, i, el)
+widenstructarray(dest::AbstractArray, i, el) = widenarray(dest, i, el)
 
 function widenarray(dest::AbstractArray{T}, i, el::S) where {S, T}
     S <: T && return dest